Convert useSkipToLastStepIfFormComplete to TS
